requirements.txt
asgiref
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
django-crispy-forms==1.9.2
djangorestframework==3.12.2
django-cors-headers==3.7.0
django-filter==2.0
django-braces==1.14.0
djangorestframework-simplejwt
django-utils-six==2.0
django-extensions
Pillow

